forced-colors-mode/forced-colors-mode-39.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Forced colors mode - alpha channel.
  Tests that the alpha channel for visited background colors are preserved in forced
  colors mode.
</title>
<link rel="help" href="https://www.w3.org/TR/css-color-adjust-1/#forced-colors-properties">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  body {
    background-color: Canvas;
  }
  a {
    background-color: rgba(255, 0, 0, .5);
  }
  a:visited {
    background-color: rgba(114, 173, 218, .3);
  }
</style>
<body id="body">
  <a href="" id="link">
    Visited background colors get their alpha channel from their non-visited
    counterpart. This alpha channel should be preserved in forced colors mode.
  </a>
</body>

<script>
  var body_color = getComputedStyle(document.querySelector('body')).backgroundColor;
  var expected_link_color = body_color.replace(')', ', 0.5)').replace('rgb(', 'rgba(');
  test(function(){
    assert_equals(getComputedStyle(link).backgroundColor, expected_link_color);
  }, "Checks that visited background color alpha channels are preserved in forced colors mode.");
</script>
